Move client classes to udmi/lib/client package
