+++ pubber/src/main/java/daq/pubber/IpProvider.java
@@ -15,7 +15,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import udmi.lib.LocalnetManagerProvider;
+import udmi.lib.client.LocalnetManagerProvider;
 import udmi.schema.FamilyLocalnetState;
 import udmi.schema.PubberConfiguration;
