+++ pubber/src/main/java/daq/pubber/MqttPublisher.java
@@ -56,7 +56,7 @@ import org.eclipse.paho.client.mqttv3.persist.MemoryPersistence;
 import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import udmi.lib.PubberHostProvider;
+import udmi.lib.client.PubberHostProvider;
 import udmi.schema.Basic;
 import udmi.schema.Config;
 import udmi.schema.EndpointConfiguration.Transport;
