+++ pubber/src/main/java/daq/pubber/Pubber.java
@@ -50,8 +50,8 @@ import java.util.function.Function;
 import org.apache.http.ConnectionClosedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import udmi.lib.DeviceManagerProvider;
-import udmi.lib.PubberHostProvider;
+import udmi.lib.client.DeviceManagerProvider;
+import udmi.lib.client.PubberHostProvider;
 import udmi.schema.Config;
 import udmi.schema.DevicePersistent;
 import udmi.schema.EndpointConfiguration;
