+++ pubber/src/main/java/daq/pubber/SystemManager.java
@@ -11,8 +11,8 @@ import java.io.PrintStream;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import udmi.lib.PubberHostProvider;
-import udmi.lib.SystemManagerProvider;
+import udmi.lib.client.PubberHostProvider;
+import udmi.lib.client.SystemManagerProvider;
 import udmi.schema.Entry;
 import udmi.schema.Level;
 import udmi.schema.Operation.SystemMode;
