+++ pubber/src/main/java/udmi/lib/client/DeviceManagerProvider.java
@@ -1,4 +1,4 @@
-package udmi.lib;
+package udmi.lib.client;
 
 import daq.pubber.FamilyProvider;
 import java.util.Map;
