+++ pubber/src/test/java/daq/pubber/PubberTest.java
@@ -15,7 +15,7 @@ import java.util.HashMap;
 import java.util.List;
 import org.junit.After;
 import org.junit.Test;
-import udmi.lib.PubberHostProvider;
+import udmi.lib.client.PubberHostProvider;
 import udmi.schema.BlobBlobsetConfig;
 import udmi.schema.BlobBlobsetConfig.BlobPhase;
 import udmi.schema.BlobsetConfig;
